Change score from uint to ulong and move controls
Change score from uint to ulong.
Move controls from PlayerController to Config.

+++ TouhouAnimismOfTheNight/Assets/Scripts/Core/GameManager.cs
@@ -19,7 +19,7 @@ namespace TH.Core
         /// <summary>
         /// The player's score
         /// </summary>
-        public uint score = 0;
+        public ulong score = 0;
 
         private void Awake()
         {
